Update event handling to use DOM 2 events
Using the older style event handling could be problematic. Instead, use
DOM 2 even handling, which will cause fewer issues.

Some of the tests had to be updated to account for the new event
handling. It was difficult to test exactly the same way, but the tests
are pretty close to what they were, except with DOM 2 event handling.